refactor: Refactor code to move GitProject and ProjectService to commons package
Renamed the gitprojects directory to commons package as it's better reflective of its content. Moved GitProject and ProjectService classes to it, now known as GitRepoHandler. All necessary import statements have also been updated to reflect this change. This refactoring improves the organization of the project and correctly categorizes common utilities.