Refactor code for AnalyzerRun and related tests
The commit primarily refactors the AnalyzerRun component and associated tests, MiningGraphQLTest, and updates how recentRuns are tested. Equality methods and hashCode functions are added or adapted for AnalyzerRun and other related entities like GitHubCommit and RemoteProject. Also, some unused imports are removed and code re-ordering is made for cleaner organization.