style: Refactor Dashboard components and filter functionality
The DashboardCard component has been deleted and its functionality has been moved into the DashBoardItem and DashBoard components. The filtering of projects on the Dashboard page is improved where now it includes both the project name and project URL. Also, minor updates were made to the Sidebar navigation menu and the layout and visuals of the Dashboard page, enhancing user experience.