feat:  Introduced CommonTypes enum
Modified the import organization in IndexOfReplaceableByContains and UnnecessaryToString classes for increased clarity. The introduction of the CommonTypes enum simplifies referencing common Java classes. This new enum was implemented in UnnecessaryToString class for more efficient type checks. Finally, the accessibility of ResultRecord class in IndexOfReplaceableByContains was reduced to private to enhance encapsulation.